backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation and settings
pydantic==2.6.1
pydantic-settings==2.1.0

# Celery and Redis
celery==5.3.6
redis==5.0.1

# ML/DL (基础依赖，可根据需要扩展)
torch==2.2.0
torchvision==0.17.0
numpy==1.26.3
Pillow==10.2.0
opencv-python==4.10.0.84
albumentations==1.3.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
black==24.1.1
isort==5.13.2
mypy==1.8.0
